# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentOS API",
    "description": "AgentOS gateway API contract.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "CompanyOwnerInputSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "input_id",
          "status",
          "version",
          "resolved_at"
        ],
        "properties": {
          "input_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNRESOLVED",
              "RESOLVED"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "resolved_at": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "CompanyOwnerInputResolutionRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "object",
            "minProperties": 1,
            "maxProperties": 128
          },
          "value_ref": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          }
        },
        "oneOf": [
          {
            "required": [
              "value"
            ],
            "not": {
              "required": [
                "value_ref"
              ]
            }
          },
          {
            "required": [
              "value_ref"
            ],
            "not": {
              "required": [
                "value"
              ]
            }
          }
        ]
      },
      "CompanyOwnerInputsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompanyOwnerInputSummary"
            }
          }
        }
      },
      "CompanyOwnerInputResolutionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "$ref": "#/components/schemas/CompanyOwnerInputSummary"
          }
        }
      },
...
      "CustomerTimelineEntry": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "occurred_at",
          "event_id",
          "event_type",
          "stage",
          "canonical_event",
          "classification"
        ],
        "properties": {
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          },
          "source_record_id": {
            "type": "string"
          },
          "event_id": {
            "type": "string"
          },
          "event_type": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "canonical_event": {
            "type": [
              "null",
              "string"
            ]
          },
          "classification": {
            "type": "string",
            "enum": [
              "FACT",
              "SIGNAL",
              "HYPOTHESIS",
              "DECISION",
              "ACTION"
            ]
          },
          "domain": {
            "type": "string",
            "enum": [
              "MARKETING",
              "SALES",
              "COMMERCE",
              "SUPPORT",
              "ORCHESTRATOR"
            ]
          },
          "summary": {
            "type": "string"
          },
          "evidence_reference": {
            "type": "string"
          },
          "gap_reason": {
            "type": "string"
          }
        }
      },
      "CustomerTimelineResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "next_cursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerTimelineEntry"
            }
          },
          "next_cursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error_code",
          "message",
          "retryable",
          "correlation_id"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "correlation_id": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/customers/{customer_id}/timeline": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "to",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "customer_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerTimelineResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/company/analytics": {
      "get": {
        "summary": "Get company analytics for a window",
        "tags": [
          "Company analytics"
        ],
        "responses": {
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/company/owner-inputs": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOwnerInputsResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/company/owner-inputs/{id}/resolve": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyOwnerInputResolutionRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "pattern": "^[A-Za-z0-9-]+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 32
            },
            "in": "header",
            "name": "if-match",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyOwnerInputResolutionResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "5XX": {
            "description": "ErrorResponse — the gateway vocabulary for a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}